+++ changelog.md
@@ -14,6 +14,7 @@ Codion Change Log
 ### is.codion.framework.domain
 - DefaultEntity.validateTypes() performance bug fixed, unnecessary HashMap lookup removed, renamed validate().
 - DefaultKey.setSerializer() and serializerForDomain() moved to EntitySerializer. 
+- SingleColumnKey added, DefaultKey renamed CompositeColumnKey.
 ### is.codion.framework.db.local
 - EntityResultPacker, minor optimization to eliminate HashMap resizing.
 ### is.codion.swing.framework.ui